Vex and RSL files compile and are useful.
The shupattern.h file now has noise mappings in it for noise, signed noise and cellnoise. RSL now has typedefs for its variable types. The surface models compile as expected in both vex and rsl. The header files for osl and glsl are simply copies of the files for rsl. For osl this is because most of the language is identical to rsl, with the notable exceptions that osl has no illuminance or gather loops (like rsl does) and rsl has no concept of a radiance closure(like osl does). I forsee little difficulty in porting code from rsl to osl.